move coupling functions in their own file
