+++ src/funcs.h
@@ -396,6 +396,7 @@ void    node_getResults(int node, double wt, float x[]);
 // --- coupling functions
 void    coupling_execute(double tStep);
 void    node_deleteOpenings(int node);
+void    node_setOldCouplingState(int j);
 
 //-----------------------------------------------------------------------------
 //   Conveyance System Inflow Methods
